Extract play music extension
